CLIP/requirements.txt
torch
clip
pillow